Using 8bit T2 instead of 16bit T1 for ModBus

+++ mymodbus.c
@@ -75,7 +75,7 @@ eMBErrorCode reg_Motor2( reg_ptr_t reg, UCHAR * pucRegBuffer, eMBRegisterMode eM
 	return reg_Motor(reg, pucRegBuffer, eMode, MOTOR2);
 }
 
-void MYMODBUS_Init(int baudrate)
+void MYMODBUS_Init(unsigned long baudrate)
 {
 	MOTOR_Init();
 	REG_Register(&r1);
