+++ mymodbus.h
@@ -8,7 +8,7 @@
 #ifndef MYMODBUS_H_
 #define MYMODBUS_H_
 
-void MYMODBUS_Init(int baudrate);
+void MYMODBUS_Init(unsigned long baudrate);
 void MYMODBUS_Manage(void);
 
 #endif /* MYMODBUS_H_ */
